mysite/requirements.txt
Django==1.9.2
django-ckeditor==5.0.3
MySQL-python==1.2.5
wheel==0.29.0
